feat(helpers): add getWorldRect
