+++ src/pointer.js
@@ -1,6 +1,6 @@
 import { getCanvas } from './core.js';
 import { on } from './events.js';
-import { getWorldRect } from './utils.js';
+import { getWorldRect } from './helpers.js';
 
 /**
  * A simple pointer API. You can use it move the main sprite or respond to a pointer event. Works with both mouse and touch events.
